services/activity-service/requirements.txt
fastapi[standard]
sqlalchemy
asyncpg
pydantic-settings
PyJWT
aio-pika
